requirements.txt
gradio
pinecone-client
openai
numpy
gspread
oauth2client